can resize speckles from keyboard

+++ src/disp/fiber_disp.h
@@ -194,7 +194,7 @@ public:
     
     
     /// size for speckle display (also know as `speckles`)
-    int speckle_size;
+    float speckle_size;
 
     /// style for speckle display (also know as `speckles[1]`)
     /**
